# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zero-Click CRM API
  version: 1.0.0
  description: >-
    Documentación de la API para el backend Zero-Click CRM. Incluye ingestión multicanal,
    procesamiento de audio/video, consultas CRM, chat asistido por IA, alertas, notas,
    knowledge base y endpoints administrativos.
  contact:
    name: Equipo Relay
servers:
  - url: http://localhost:4000
    description: Servidor de desarrollo local
  - url: https://your-deployment-url.example
    description: Servidor de producción (actualiza esta URL al desplegar)
...
paths:
...
  /api/knowledge:
    get:
      summary: Listar entradas de la base de conocimiento
      tags: [Knowledge]
      parameters:
        - $ref: '#/components/parameters/CompanyIdQuery'
        - name: search
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Entradas paginadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeListResponse'
    post:
      summary: Crear entrada de conocimiento
      tags: [Knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeCreateRequest'
      responses:
        '200':
          description: Entrada creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeEntryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/knowledge/upload:
    post:
      summary: Crear múltiples entradas desde texto/base64
      tags: [Knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeUploadRequest'
      responses:
        '200':
          description: Entradas creadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/KnowledgeEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/knowledge/search:
    post:
      summary: Buscar en knowledge con embeddings (si disponibles)
      tags: [Knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeSearchRequest'
      responses:
        '200':
          description: Resultados por similitud o búsqueda simple
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KnowledgeSearchResult'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/knowledge/{id}:
    delete:
      summary: Eliminar entrada de conocimiento por ID
      tags: [Knowledge]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entrada eliminada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    KnowledgeEntry:
      type: object
      properties:
        id:
          type: string
        company_id:
          type: string
          nullable: true
        title:
          type: string
        content:
          type: string
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
    KnowledgeListResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeEntry'
        count:
          type: integer
    KnowledgeCreateRequest:
      type: object
      required: [content]
      properties:
        title:
          type: string
          nullable: true
        content:
          type: string
        companyId:
          type: string
          nullable: true
        metadata:
          type: object
          default: {}
    KnowledgeEntryResponse:
      type: object
      properties:
        ok:
          type: boolean
        entry:
          $ref: '#/components/schemas/KnowledgeEntry'
    KnowledgeUploadRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        fileBase64:
          type: string
        chunkSize:
          type: integer
          default: 1600
        companyId:
          type: string
          nullable: true
        metadata:
          type: object
          default: {}
    KnowledgeSearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        companyId:
          type: string
          nullable: true
        limit:
          type: integer
          default: 5
    KnowledgeSearchResult:
      type: object
      properties:
        id:
          type: string
        sourceId:
          type: string
        companyId:
          type: string
          nullable: true
        contactId:
          type: string
          nullable: true
        text:
          type: string
        metadata:
          type: object
        similarity:
          type: number
        created_at:
          type: string
          format: date-time
...
  parameters:
    CompanyIdQuery:
      name: companyId
      in: query
      schema:
        type: string
      description: Filtra por UUID de la compañía
    LimitQuery:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
    OffsetQuery:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
...
  responses:
    BadRequest:
      description: Petición inválida
      content:
        application/json:
          schema:
            type: object
            properties:
              ok:
                type: boolean
                example: false
              error:
                type: string
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              ok:
                type: boolean
                example: false
              error:
                type: string
...